# src/ModelLayer/DTO/UserRegisterDTO.cs
namespace ModelLayer
{
    public class UserRegisterDTO
    {
        public string FullName { get; set; } = string.Empty;
        public string Email { get; set; } = string.Empty;
        public string Password { get; set; } = string.Empty;
        public string Role { get; set; } = "User";
    }
}