# swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "QuantityMeasurementWebApi",
    "version": "1.0"
  },
  "paths": {
    "/api/v1/auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRegisterDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRegisterDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UserRegisterDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserLoginDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UserLoginDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UserLoginDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/quantities/compare": {
      "post": {
        "tags": [
          "QuantityMeasurement"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuantityInputDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/QuantityInputDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/QuantityInputDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/quantities/convert": {
      "post": {
        "tags": [
          "QuantityMeasurement"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuantityInputDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/QuantityInputDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/QuantityInputDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/quantities/add": {
      "post": {
        "tags": [
          "QuantityMeasurement"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuantityInputDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/QuantityInputDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/QuantityInputDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/quantities/subtract": {
      "post": {
        "tags": [
          "QuantityMeasurement"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuantityInputDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/QuantityInputDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/QuantityInputDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/quantities/divide": {
      "post": {
        "tags": [
          "QuantityMeasurement"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuantityInputDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/QuantityInputDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/QuantityInputDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/quantities/history/operation/{operationType}": {
      "get": {
        "tags": [
          "QuantityMeasurement"
        ],
        "parameters": [
          {
            "name": "operationType",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/quantities/history/type/{measurementType}": {
      "get": {
        "tags": [
          "QuantityMeasurement"
        ],
        "parameters": [
          {
            "name": "measurementType",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/quantities/count": {
      "get": {
        "tags": [
          "QuantityMeasurement"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "QuantityDTO": {
        "required": [
          "measurementType",
          "unit",
          "value"
        ],
        "type": "object",
        "properties": {
          "value": {
            "type": "number",
            "format": "double"
          },
          "unit": {
            "minLength": 1,
            "type": "string"
          },
          "measurementType": {
            "minLength": 1,
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "QuantityInputDTO": {
        "required": [
          "thatQuantityDTO",
          "thisQuantityDTO"
        ],
        "type": "object",
        "properties": {
          "thisQuantityDTO": {
            "$ref": "#/components/schemas/QuantityDTO"
          },
          "thatQuantityDTO": {
            "$ref": "#/components/schemas/QuantityDTO"
          },
          "targetQuantityDTO": {
            "$ref": "#/components/schemas/QuantityDTO"
          }
        },
        "additionalProperties": false
      },
      "UserLoginDTO": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "nullable": true
          },
          "password": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "UserRegisterDTO": {
        "type": "object",
        "properties": {
          "fullName": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "password": {
            "type": "string",
            "nullable": true
          },
          "role": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "JWT Authorization header using the Bearer scheme. Enter 'Bearer' followed by your token. Example: Bearer eyJhbGciOi...",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "components/securitySchemes#/components/securitySchemes/Bearer": [ ]
    }
  ],
  "tags": [
    {
      "name": "Auth"
    },
    {
      "name": "QuantityMeasurement"
    }
  ]
}